removed redundant task3 codes
